Added default value ForceAttemptHTTP2 in transport

+++ exporter/signalfxexporter/config.go
@@ -106,7 +106,7 @@ type Config struct {
 	NonAlphanumericDimensionChars string `mapstructure:"nonalphanumeric_dimension_chars"`
 
 	// MaxConnections is used to set a limit to the maximum idle HTTP connection the exporter can keep open.
-	MaxConnections uint `mapstructure:"max_connections"`
+	MaxConnections int64 `mapstructure:"max_connections"`
 }
 
 func (cfg *Config) getOptionsFromConfig() (*exporterOptions, error) {
